Se agrega scheduler para modo expropiativo

+++ Progra1/Taylor.h
@@ -1,7 +1,7 @@
 #ifndef TAYLOR_H
 #define TAYLOR_H
 
-long double arcenesimo (int n);
+long double arcenesimo (long n);
 long double sumarN (int n);
 
 #endif
